deployment_hf/requirements.txt
ultralytics
gradio>=4.44.1
opencv-python-headless
pillow
numpy
huggingface-hub<0.27.0
pydantic>=2.0,<2.10
